Storing the LAN where an achievement was awarded

+++ app/Http/Controllers/UserAchievementController.php
@@ -3,9 +3,11 @@
 namespace Zeropingheroes\Lanager\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Zeropingheroes\Lanager\Achievement;
+use Zeropingheroes\Lanager\Lan;
 use Zeropingheroes\Lanager\Requests\DestroyUserAchievementRequest;
+use Zeropingheroes\Lanager\Requests\StoreUserAchievementRequest;
 use Zeropingheroes\Lanager\UserAchievement;
-use Zeropingheroes\Lanager\User;
 use Illuminate\Support\Facades\View;
 
 class UserAchievementController extends Controller
@@ -13,17 +15,23 @@ class UserAchievementController extends Controller
     /**
      * Display a listing of the resource.
      *
+     * @param Lan $lan
      * @return \Illuminate\Contracts\View\View
      */
-    public function index()
+    public function index(Lan $lan)
     {
         $this->authorize('index', UserAchievement::class);
 
-        $userAchievements = UserAchievement::with('user', 'achievement')->get();
-        $users = User::orderBy('username')->get();
+        $userAchievements = $lan->userAchievements()
+            ->with('user', 'achievement')
+            ->get();
+        $users = $lan->users()
+            ->orderBy('username')
+            ->get();
         $achievements = Achievement::all();
 
         return View::make('pages.user-achievements.index')
+            ->with('lan', $lan)
             ->with('userAchievements', $userAchievements)
             ->with('users', $users)
             ->with('achievements', $achievements);
@@ -33,14 +41,16 @@ class UserAchievementController extends Controller
      * Store a newly created resource in storage.
      *
      * @param Request $httpRequest
+     * @param Lan $lan
      * @return \Illuminate\Http\Response
      * @internal param Request|StoreUserAchievementRequest $request
      */
-    public function store(Request $httpRequest)
+    public function store(Request $httpRequest, Lan $lan)
     {
         $this->authorize('create', UserAchievement::class);
 
         $input = $httpRequest->only(['user_id', 'achievement_id']);
+        $input['lan_id'] = $lan->id;
 
         $request = new StoreUserAchievementRequest($input);
 
@@ -54,7 +64,7 @@ class UserAchievementController extends Controller
         $userAchievement = UserAchievement::create($input);
 
         return redirect()
-            ->route('user-achievements.index')
+            ->route('lans.user-achievements.index', $lan)
             ->withSuccess(
                 __(
                     'phrase.achievement-successfully-awarded',
@@ -66,17 +76,23 @@ class UserAchievementController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param Lan $lan
      * @param  \Zeropingheroes\Lanager\UserAchievement $userAchievement
      * @return \Illuminate\Http\Response
      */
-    public function destroy(UserAchievement $userAchievement)
+    public function destroy(Lan $lan, UserAchievement $userAchievement)
     {
         $this->authorize('delete', $userAchievement);
+
+        // If the user achievement is accessed via the wrong LAN ID, show 404
+        if ($userAchievement->lan_id != $lan->id) {
+            abort(404);
+        }
 
         UserAchievement::destroy($userAchievement->id);
 
         return redirect()
-            ->route('user-achievements.index')
+            ->route('lans.user-achievements.index', $lan)
             ->withSuccess(
                 __(
                     'phrase.achievement-successfully-revoked',
